+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 This is the rolling changelog for TShock for Terraria. Use past tense when adding new entries; sign your name off when you add or change something. This should primarily be things like user changes, not necessarily codebase changes unless it's really relevant or large.
 
+## Upcoming changes
+* Removed `/confuse` command and Terraria player data resync from @Zidonuke. (@hakusaro)
+
 ## TShock 4.4.0 (Pre-release 3)
 * Fixed `/worldmode` command to correctly target world mode. (@Ristellise)
 * The following commands have been removed: `tbloodmoon`, `invade`, `dropmeteor`. `fullmoon`, `sandstorm`, `rain`, `eclipse`